Link the Shell interop assembly and move the selector to the ProjectSystem namespace.
