+++ src/Microsoft.VisualStudio.ProjectSystem.FSharp.VS/ProjectSystem/VS/FSharpProjectSelector.cs
@@ -1,29 +1,34 @@
 ﻿using System;
-using Microsoft.VisualStudio.Shell.Interop1;
 using System.Xml.Linq;
 using System.Xml.XPath;
 using System.Runtime.InteropServices;
 using System.Xml;
+using Microsoft.VisualStudio.Packaging;
 using Microsoft.VisualStudio.ProjectSystem.VS.Generators;
+using Microsoft.VisualStudio.Shell.Interop;
 
-namespace Microsoft.VisualStudio.Packaging
+namespace Microsoft.VisualStudio.ProjectSystem.VS
 {
     [ComVisible(true)]
     [Guid(SelectorGuid)]
-    [ClassRegistration(SelectorGuid, "Microsoft.VisualStudio.Packaging.FSharpProjectSelector, Microsoft.VisualStudio.ProjectSystem.FSharp.VS")]
+    [ClassRegistration(SelectorGuid, AssemblyQualifiedSelectorTypeName)]
     public sealed class FSharpProjectSelector : IVsProjectSelector
     {
         public const string SelectorGuid = "E720DAD0-1854-47FC-93AF-E719B54B84E6";
+        public const string AssemblyQualifiedSelectorTypeName = "Microsoft.VisualStudio.ProjectSystem.VS.FSharpProjectSelector, Microsoft.VisualStudio.ProjectSystem.FSharp.VS";
 
         public void GetProjectFactoryGuid(Guid guidProjectType, string pszFilename, out Guid guidProjectFactory)
         {
             XDocument doc = XDocument.Load(pszFilename);
             XmlNamespaceManager nsm = new XmlNamespaceManager(new NameTable());
             nsm.AddNamespace("msb", "http://schemas.microsoft.com/developer/msbuild/2003");
-            var hasProjectElementWithSdkAttribute = doc.XPathSelectElement("/msb:Project[@Sdk]", nsm) != null || doc.XPathSelectElement("/Project[@Sdk]") != null;
-            var hasImportElementWithSdk = doc.XPathSelectElement("msb:Import[@Sdk]", nsm) != null || doc.XPathSelectElement("Import[@Sdk]") != null;
 
-            if (hasProjectElementWithSdkAttribute || hasImportElementWithSdk)
+            // If the project has either a Project-level SDK attribute or an Import-level SDK attribute, we'll open it with the new project system.
+            // Make sure to check for both namespace-qualified and without for projects that get rid of the xmlns attribute.
+            var hasProjectElementWithSdkAttribute = doc.XPathSelectElement("/msb:Project[@Sdk]", nsm) != null || doc.XPathSelectElement("/Project[@Sdk]") != null;
+            var hasImportElementWithSdkAttribute = doc.XPathSelectElement("msb:Import[@Sdk]", nsm) != null || doc.XPathSelectElement("Import[@Sdk]") != null;
+
+            if (hasProjectElementWithSdkAttribute || hasImportElementWithSdkAttribute)
             {
                 guidProjectFactory = Guid.Parse(FSharpProjectSystemPackage.ProjectTypeGuid);
                 return;
